warnings bad
git-svn-id: http://code.icecube.wisc.edu/svn/projects/dataclasses/trunk@36017 16731396-06f5-0310-8873-f7f720988828

+++ private/test/I3TreeUtilsTest.cxx
@@ -23,10 +23,10 @@ TEST_GROUP(I3TreeUtilsTest);
 class int_t
 {
 public:
-  int id;
+  uint64_t id;
   const int GetMinorID() const { return id; }
-  const int GetMajorID() const { return id; }
-  int_t(int i) : id(i) { }
+  const uint64_t GetMajorID() const { return id; }
+  int_t(uint64_t i) : id(i) { }
 };
 
 TEST(a_insert_and_append)
